requirements.txt
streamlit
langchain
langchain-groq
langchain-community
python-dotenv
numexpr